[Fix] ChidlrenController -> UserController로 변경
