--- src/main/java/com/hana4/sonjumoney/controller/ChildController.java
@@ -1,32 +1,0 @@
-package com.hana4.sonjumoney.controller;
-
-import java.util.List;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.hana4.sonjumoney.dto.response.GetChildrenResponse;
-import com.hana4.sonjumoney.service.ChildService;
-import com.hana4.sonjumoney.util.AuthenticationUtil;
-
-import io.swagger.v3.oas.annotations.tags.Tag;
-import lombok.RequiredArgsConstructor;
-
-@Tag(name = "Children", description = "내 아이 관련 API")
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/children")
-public class ChildController {
-
-	private final ChildService childService;
-
-	@GetMapping()
-	public ResponseEntity<List<GetChildrenResponse>> getChildren(Authentication authentication) {
-		Long userId = AuthenticationUtil.getUserId(authentication);
-		return ResponseEntity.ok().body(childService.getChildren(userId));
-	}
-
-}
