+++ src/main/java/com/hana4/sonjumoney/controller/UserController.java
@@ -1,11 +1,14 @@
 package com.hana4.sonjumoney.controller;
 
+import java.util.List;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.hana4.sonjumoney.dto.response.GetChildrenResponse;
 import com.hana4.sonjumoney.dto.response.UserInfoResponse;
 import com.hana4.sonjumoney.service.UserService;
 import com.hana4.sonjumoney.util.AuthenticationUtil;
@@ -23,4 +26,11 @@ public class UserController {
 		System.out.println("userinfo 컨르롤러 진입");
 		return ResponseEntity.ok().body(userService.getUserByUserId(AuthenticationUtil.getUserId(authentication)));
 	}
+
+	@GetMapping("/children")
+	public ResponseEntity<List<GetChildrenResponse>> getChildren(Authentication authentication) {
+		Long userId = AuthenticationUtil.getUserId(authentication);
+		return ResponseEntity.ok().body(userService.getChildren(userId));
+	}
+
 }
