--- src/test/java/com/hana4/sonjumoney/controller/ChildControllerTest.java
@@ -1,40 +1,0 @@
-package com.hana4.sonjumoney.controller;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.hana4.sonjumoney.ControllerTest;
-
-@SpringBootTest
-@AutoConfigureMockMvc
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-public class ChildControllerTest extends ControllerTest {
-	@Autowired
-	MockMvc mockMvc;
-
-	@Autowired
-	ObjectMapper objectMapper;
-
-	@Test
-	@DisplayName("내 아이 조회 테스트")
-	void getChildrenTest() throws Exception {
-		mockMvc.perform(get("/api/children")
-				.header("Authorization", "Bearer " + accessToken)
-				.contentType(MediaType.APPLICATION_JSON))
-			.andExpect(status().isOk())
-			.andExpect(jsonPath("$[0].user_id").value(3))
-			.andExpect(jsonPath("$[0].user_name").value("용돈 받을 애"))
-			.andDo(print());
-	}
-}
