+++ src/test/java/com/hana4/sonjumoney/controller/UserControllerTest.java
@@ -1,7 +1,7 @@
 package com.hana4.sonjumoney.controller;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import java.io.IOException;
@@ -41,7 +41,6 @@ public class UserControllerTest {
 	@Autowired
 	private UserService userService;
 
-
 	private static String accessToken;
 
 	@BeforeAll
@@ -54,27 +53,39 @@ public class UserControllerTest {
 				.andExpect(status().isOk())
 				.andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON));
 			JwtTokenDto jwtTokenDto = objectMapper.readValue(
-				resultActions.andReturn().getResponse().getContentAsString(),JwtTokenDto.class);
-			accessToken= jwtTokenDto.access_token();
+				resultActions.andReturn().getResponse().getContentAsString(), JwtTokenDto.class);
+			accessToken = jwtTokenDto.access_token();
 		} catch (Exception e) {
 			throw new RuntimeException(e);
 		}
 	}
 
 	@AfterAll
-	public void cleanUp()throws IOException {
+	public void cleanUp() throws IOException {
 
 	}
 
 	@Test
 	@DisplayName("/user - User info test")
-	void getUserInfo()throws Exception {
+	void getUserInfo() throws Exception {
 		String url = "/api/users";
 		mockMvc.perform(
 				get(url)
 					.header("Authorization", "Bearer " + accessToken)
 					.contentType(MediaType.APPLICATION_JSON))
 			.andExpect(status().isOk());
+	}
+
+	@Test
+	@DisplayName("내 아이 조회 테스트")
+	void getChildrenTest() throws Exception {
+		mockMvc.perform(get("/api/users/children")
+				.header("Authorization", "Bearer " + accessToken)
+				.contentType(MediaType.APPLICATION_JSON))
+			.andExpect(status().isOk())
+			.andExpect(jsonPath("$[0].user_id").value(3))
+			.andExpect(jsonPath("$[0].user_name").value("용돈 받을 애"))
+			.andDo(print());
 	}
 
 }
